# application/services/custom_services/monkchartofaccounts/src/models/entitymodels/account-info.ts
export interface account-info 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   account-name: String,
   account-id: String,
   description: String,
   tags: String,
   account-storage-created: Boolean
}